ncursesw
so I can build ncurses with widechar support, but that results in
creating ncursesw as the header folder, which no one will use. I disable
the w suffix when building ncurses, but nano is adamant that it wants to
try linking libncursesw.a. So I add a dummy libncursesw.a.