save config.cache for each package build
earlier we were using a more minimal config.cache file with common
across all builds, but that did not improve build times as much. Now we
use the file saved_config.cache stored along with the superconfigure
file, that is unique to each package build.

It's a bit risky, because we might cause some errors that could
otherwise be avoided. We might also have a bit of an issue when updating
the deps for a package or adding a new package (the copycache and
cleancache scripts might help with this).

Let's see how much it improves build times.