upgrade emacs to 29.3 with jansson
I disabled the check that emacs has at the end of the build, because
although it gives the right output on the terminal, for some reason
(only on aarch64), the makefile seems to think it doesn't. If someone
can figure why this build error is happening, that would be nice, and we
can remove the patch to the Makefile.