check sha256sum after downloading
this PR adds a new step to the build process of superconfigure --
checking the SHA256 checksum of the downloaded source tarball.

After downloading the source tarball and creating o/%/downloaded, the
next step is to check the SHA256 checksum of the downloaded file with a
checksum file that is part of the superconfigure repo. The checksum file
is called check.signature, and needs to be manually added along with
every new BUILD.mk. The tarball is unzipped only if the checksums match.

If we're downloading from a git repo or have some other setup there, you
can skip the checksum process just like you skip the downloading and/or
patching process. But you still need to create the check.signature file,
otherwise the makefile won't work.