avoid creating /zip by using renamestr
https://github.com/jart/cosmopolitan/pull/1344

the renamestr program written in the above PR means the build process
for binaries in superconfigure does not require the existence of /zip
or equivalent folder as an absolute path to copy necessary assets. we
build the binaries setting $COSMOS as the target folder for all files
and then patch the .rodata section in the debug binaries right before
running apelink (actually, we run apelink with copies of the binaries
so that the originals can be used like regular ELF binaries installed
on a traditional Linux system.) Now, assuming binfmt_misc details are
satisfied, we should be able to build software without needing sudo.

/zip was needed in the build configuration of packages because it was
(at the time) the only way to ensure the final APE binary had correct
locations for reading assets. But /zip had several side effects: need
of sudo for creating the folder, COSMOPOLITAN_DISABLE_ZIPOS=1 for all
steps of superconfigure, debug binaries were difficult to use in case
we needed to fix a bug, some build flakes due to builds writing files
to the same subfolder of /zip, et cetera. Now, the build procedure of
superconfigure should (ideally) result in ELF binaries installed from
using cosmocc, and the /zip features are activated right when the fat
binary is created, which means finding errors is far more streamlined