minimize debug output
the individual commands of superconfigure have lot of debugging output,
which makes it difficult to track the current builds or failures. With
this commit the default behavior is to pipe stdout/stderr output to a
specified NULLFILE (set as /dev/null in variables.mk), and only echo a
line for each step. the debug can be redirected to a file via setting

make LOG=file o/your/target/built.x86_64

the output will be in 2 files created at each step (a .log and a .err)

or you can set it to the old behavior of viewing all output in the cli:

make LOG=stdout o/your/target/built.x86_64

to add this to any custom step in the makefiles, use $(ERRLOG).